Add notes list and editor components
The note-list component renders the notes and make each of them a
link to the note.

The height of the list and the editor is set to 95vh, a toolbar
will be added at the top in the future to create and delete notes.

Use overflow-y: auto to let the note list be scrollable.

The title and the content of the list are disabled in the list and
shortened if the text is too long. Use text-overflow: ellipsis to
add `...`.