Handle Create and Delete
Add a navbar with buttons to create and delete notes.
The first toolbar is above the list and doesn't need to have a
note selected to show.
The second toolbar is above the editor and will show when a note
is activated: The commands to edit the selected note will be
located there.

The create button is disabled when the last note created is empty.
This solves 2 problems: Being able to click Create 10 times and
rendering an empty list (no title or content). We will be able to
use the next autoincrement to directly select the new note.

The delete button uses window.confirm to let the user confirm the
irreversible action.